requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
PyQt5>=5.15.0
numpy>=1.24.0
PyYAML>=6.0
supervision>=0.18.0
lap>=0.4.0
pandas   # for csv export
